+++ apps/nowcasting-app/pages/index.tsx
@@ -3,7 +3,7 @@ import { withPageAuthRequired } from "@auth0/nextjs-auth0";
 import Layout from "../components/layout";
 import { PvLatestMap } from "../components/map";
 import SideLayout from "../components/side-layout";
-import PvRemixChart from "../components/charts/pvRemixChart";
+import PvRemixChart from "../components/charts/pv-remix-chart";
 
 export default function Home() {
   return (
